refactor: separate ARK router
